requirements.txt
google.generativeai
requests
loging
PySide6
python-telegram-bot